+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - Fix create user by API [#2609](https://github.com/opendatateam/udata/pull/2609)
 - Add sqlite, db and ics to allowed extensions [#2610](https://github.com/opendatateam/udata/pull/2610)
+- Better markup parsing [#2611](https://github.com/opendatateam/udata/pull/2611):
+  - Geozone's and Resource type's labelize function return None if no object is found.
+  - New SafeMarkup class, which inherits from Markup, uses Bleach to sanitize Markup class.
 
 ## 2.6.4 (2021-03-24)